# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabShop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl?: string | null;
  productCount: number;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  description: string;
  imageUrl?: string | null;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  sku: string;
  stockQty: number;
  lowStockThreshold: number;
  imageUrl?: string | null;
  featured: boolean;
  categoryId?: number | null;
  categoryName?: string | null;
  specs?: string | null;
}

export interface CreateProductBody {
  name: string;
  slug: string;
  description: string;
  price: number;
  sku: string;
  stockQty: number;
  lowStockThreshold: number;
  imageUrl?: string | null;
  featured: boolean;
  categoryId?: number | null;
  specs?: string | null;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  subtotal: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  sessionId: string;
  status: OrderStatus;
  total: number;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  clerkUserId?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export interface CreateOrderBody {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  inStockCount: number;
}

export interface CategoryCount {
  categoryId: number;
  categoryName: string;
  count: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  sku: string;
  stockQty: number;
  lowStockThreshold: number;
  isLowStock: boolean;
  price: number;
  categoryName?: string | null;
  imageUrl?: string | null;
}

export interface AdjustStockBody {
  /** Positive to add, negative to remove */
  adjustment: number;
  reason?: string | null;
}

export interface AdminDashboard {
  totalProducts: number;
  totalCategories: number;
  totalOrders: number;
  totalRevenue: number;
  lowStockCount: number;
  recentOrders: Order[];
  lowStockItems: InventoryItem[];
}

export interface SendAlertBody {
  email: string;
  productIds?: number[] | null;
}

export interface AlertResult {
  sent: boolean;
  message: string;
  itemCount: number;
}

export interface ProductLabel {
  id: number;
  name: string;
  sku: string;
  price: number;
  barcode: string;
  categoryName?: string | null;
  stockQty: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  inStock?: boolean;
  featured?: boolean;
};

export type GetCartParams = {
  sessionId: string;
};