# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FabShop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add an item to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/item/{cartItemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove an item from the cart
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /admin/inventory/{id}/adjust:
    post:
      operationId: adjustStock
      tags: [admin]
      summary: Adjust stock quantity for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustStockBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /admin/send-low-stock-alert:
    post:
      operationId: sendLowStockAlert
      tags: [admin]
      summary: Send low stock email alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAlertBody"
      responses:
        "200":
          description: Alert sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertResult"
  /admin/label/{productId}:
    get:
      operationId: getProductLabel
      tags: [admin]
      summary: Get label data for a product (for printing)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Label data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductLabel"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - productCount
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
      required:
        - name
        - slug
        - description
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
          nullable: true
        price:
          type: number
          format: float
        quantity:
          type: integer
        subtotal:
          type: number
          format: float
      required:
        - id
        - productId
        - productName
        - price
        - quantity
        - subtotal
    Cart:
      type: object
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
          format: float
        itemCount:
          type: integer
      required:
        - sessionId
        - items
        - total
        - itemCount
    AddToCartBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - sessionId
        - productId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        stockQty:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        price:
          type: number
          format: float
        categoryName:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - id
        - name
        - sku
        - stockQty
        - lowStockThreshold
        - isLowStock
        - price
    AdjustStockBody:
      type: object
      properties:
        adjustment:
          type: integer
          description: Positive to add, negative to remove
        reason:
          type: string
          nullable: true
      required:
        - adjustment
...
    SendAlertBody:
      type: object
      properties:
        email:
          type: string
        productIds:
          type: array
          items:
            type: integer
          nullable: true
      required:
        - email
    AlertResult:
      type: object
      properties:
        sent:
          type: boolean
        message:
          type: string
        itemCount:
          type: integer
      required:
        - sent
        - message
        - itemCount
    ProductLabel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        price:
          type: number
          format: float
        barcode:
          type: string
        categoryName:
          type: string
          nullable: true
        stockQty:
          type: integer
      required:
        - id
        - name
        - sku
        - price
        - barcode
        - stockQty