# lib/api-zod/src/generated/types/adjustStockBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabShop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdjustStockBody {
  /** Positive to add, negative to remove */
  adjustment: number;
  reason?: string | null;
}