# lib/api-zod/src/generated/types/productLabel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabShop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductLabel {
  id: number;
  name: string;
  sku: string;
  price: number;
  barcode: string;
  categoryName?: string | null;
  stockQty: number;
}